+++ src/transformers/index.js
@@ -1,20 +1,20 @@
-const inline = require('./inline')
+const inline = require('./inlineCss')
 const minify = require('./minify')
 const markdown = require('./markdown')
 const prettify = require('./prettify')
-const ensureSixHEX = require('./six-hex')
-const addURLParams = require('./url-params')
+const ensureSixHEX = require('./sixHex')
+const addURLParams = require('./urlParameters')
 const transformContents = require('./transform')
-const preventWidows = require('./prevent-widows')
-const replaceStrings = require('./replace-strings')
-const safeClassNames = require('./safe-class-names')
-const applyBaseImageUrl = require('./base-image-url')
-const removeUnusedCSS = require('./remove-unused-css')
-const removeAttributes = require('./remove-attributes')
-const attributeToStyle = require('./attribute-to-style')
-const removeInlineSizes = require('./remove-inline-sizes')
-const applyExtraAttributes = require('./extra-attributes')
-const removeInlineBgColor = require('./remove-inline-bgcolor')
+const preventWidows = require('./preventWidows')
+const replaceStrings = require('./replaceStrings')
+const safeClassNames = require('./safeClassNames')
+const applyBaseImageUrl = require('./baseImageUrl')
+const removeUnusedCSS = require('./removeUnusedCss')
+const removeAttributes = require('./removeAttributes')
+const attributeToStyle = require('./attributeToStyle')
+const removeInlineSizes = require('./removeInlineSizes')
+const applyExtraAttributes = require('./extraAttributes')
+const removeInlineBgColor = require('./removeInlineBackgroundColor')
 
 exports.process = async (html, config) => {
   html = await safeClassNames(html, config)
