+++ src/transformers/removeAttributes.js
