+++ src/transformers/removeInlineBackgroundColor.js
