+++ src/transformers/removeInlineSizes.js
