+++ src/transformers/safeClassNames.js
