+++ src/transformers/sixHex.js
