+++ xo.config.js
@@ -8,6 +8,7 @@ module.exports = {
     'import/extensions': 0,
     'operator-linebreak': 0,
     'max-nested-callbacks': 0,
+    'unicorn/filename-case': 0,
     'unicorn/prefer-ternary': 0,
     'unicorn/string-content': 0,
     'unicorn/no-array-for-each': 0,
